Former-commit-id: 43edf92b0c27293a179bfec3b5f75ad9d61cc61c
Former-commit-id: 5530830717e183d621534d57a8689aa6bb270464

+++ include/mat_vec_types.h
@@ -10,10 +10,10 @@
  */
 #ifdef USE_DOUBLE
 typedef double scalar;
-typedef long double geoscalar;
+typedef double geoscalar;
 #else
 typedef float scalar;
-typedef long double geoscalar;
+typedef double geoscalar;
 #endif 
 //typedef long double scalar;
